Add units of measure to var names
This makes the code easily readable to know that the various
calculations are using numbers in the appropriate units of measure.
The only functional change here is that in 2 places a vr was being used
firstly in msec then *1000 to be in usec, so I made 2 vars for that.
Does this make the code more maintainable for future readers?